ACDEV-1765: Transform integration test into unit test
This test was passing in Java 7 but failing in Java 8. That's because
it was an integration test, and deep inside Nimbus, the JSON serialisation
to String depends on the order of a Set, which is not guaranteed. A
change in the order will also change the signature of the token, making
the test fail.

In general, the same payload can be serialised as different equivalent
but not equal Strings, which led to completely different signatures.
Therefore, we cannot reliably compare the actual signed token with the
the expected one.

I transformed the test into a unit test. We assume that Nimbus does its
job and signs the tokens correctly.