Reverting percent encoding patch.
This turns out to be a little more squirrelly that first thought. What is actually
happening is that URLEncoder is converting space to '+', and then JWT is converting
it back again. This is seemingly different from what the spec says, but this is out
there now so should probably remain.

Spec here: https://developer.atlassian.com/static/connect/docs/latest/concepts/understanding-jwt.html